python: embedding generate_sim_solver in class constructor
